+++ CHANGELOG.md
@@ -12,6 +12,9 @@ You can also check the
 # Unreleased
 
 - Features
+  - Interactive filters area is now open by default
+  - When interactive filters area is initialized open, it's no longer possible
+    to collapse it
   - Valid cube iris in Statistics page (coming from the same data source as the
     environment's default one) are now pointing to Visualize dataset previews
 - Maintenance